@@ -73,7 +73,7 @@ ActiveAdmin.register Group do
       redirect_to action: :show
     else
       flash[:error] = @group.errors.full_messages
-      redirect_to :back
+      redirect_back(fallback_location: admin_root_path)
     end
   end
 end
