@@ -1,15 +1,5 @@
-require File.expand_path('../boot', __FILE__)
-
-require 'rails'
-# Pick the frameworks you want:
-require 'active_model/railtie'
-require 'active_job/railtie'
-require 'active_record/railtie'
-require 'action_controller/railtie'
-require 'action_mailer/railtie'
-require 'action_view/railtie'
-require 'sprockets/railtie'
-# require "rails/test_unit/railtie"
+require_relative 'boot'
+require 'rails/all'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -34,9 +24,6 @@ module Dash
 
     # Add custom validators path
     config.autoload_paths += Dir["#{config.root}/app/validators"]
-
-    # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
 
     # Setting Generator command
     config.generators do |g|
