@@ -11,7 +11,7 @@ describe 'Admin::HelpText', type: :feature do
     fill_in '名前', with: new_group.name
     fill_in 'メールアドレス', with: new_group.email
     fill_in '説明', with: new_group.description
-    click_on 'Create グループ'
+    click_on 'グループを作成'
   end
 
   describe '#index' do
